Only support bazelmod and remove aspect skylib dependency (#9)
Due to the release of Bazel 8 making Bazelmod default, i'm removing
support for Bazel repositories that use WORKSPACE (previous releases
will still work) to make this a little simpler to maintain and
understand.

I've also removed the aspect skylib dependency which was used to convert
bash into Windows Batch commands in favour of sticking with the
recommended way of supporting Bazel on Windows with MSYS2 which includes
`bash`.